Add files via upload
Datasets with text fragments from astrophysics papers, provided by the NASA Astrophysical Data System with manually tagged astronomical facilities and other entities of interest (e.g., celestial objects).
Datasets are in JSON Lines format (each line is a json dictionary).
The datasets are formatted similarly to the CONLL2003 format. Each token is associated with an NER tag. The tags follow the "B-" and "I-" convention from the IOB2 syntax

Each entry consists of a dictionary with the following keys:
"unique_id": a unique identifier for this data sample. Must be included in the predictions.
"tokens": the list of tokens (strings) that form the text of this sample. Must be included in the predictions.
"ner_tags": the list of NER tags (in IOB2 format)